+++ lib/config/routes/routes.dart
@@ -1,23 +1,23 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:itacov/constant/constant.dart';
-import 'package:itacov/view/do_page.dart';
-import 'package:itacov/view/home_screen.dart';
-import 'package:itacov/view/main_page.dart';
-import 'package:itacov/view/splash.dart';
-import 'package:itacov/view/news_page.dart';
+import 'package:itacov/ui/constant/constant.dart';
+import 'package:itacov/ui/view/do_page.dart';
+import 'package:itacov/ui/view/home_page.dart';
+import 'package:itacov/ui/view/main_page.dart';
+import 'package:itacov/ui/view/news_page.dart';
+import 'package:itacov/ui/view/splash_page.dart';
 
 class Routes {
   Route<dynamic> onGenerateRoute(RouteSettings settings) {
     switch (settings.name) {
       case splashPage:
-        return MaterialPageRoute(builder: (_) => Splash());
+        return MaterialPageRoute(builder: (_) => SplashPage());
         break;
       case mainPage:
         return MaterialPageRoute(builder: (_) => MainPage());
         break;
       case homePage:
-        return MaterialPageRoute(builder: (_) => HomeScreen());
+        return MaterialPageRoute(builder: (_) => HomePage());
         break;
       case doPage:
         return MaterialPageRoute(builder: (_) => DoPage());
