+++ lib/core/bloc/corona_bloc.dart
