+++ lib/core/model/corona_model.dart
