+++ lib/ui/shared/images.dart
@@ -1,0 +1,2 @@
+
+const backgroundImage = 'assets/images/bg.png';
