+++ lib/ui/view/splash_page.dart
@@ -1,16 +1,16 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:itacov/constant/constant.dart';
+import 'package:itacov/ui/constant/constant.dart';
 
-class Splash extends StatefulWidget {
-  const Splash({Key key}) : super(key: key);
+class SplashPage extends StatefulWidget {
+  const SplashPage({Key key}) : super(key: key);
 
   @override
-  _SplashState createState() => _SplashState();
+  _SplashPageState createState() => _SplashPageState();
 }
 
-class _SplashState extends State<Splash> {
+class _SplashPageState extends State<SplashPage> {
   @override
   void initState() {
     super.initState();
