+++ lib/ui/widgets/body_widgets.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:itacov/constant/constant.dart';
-import 'package:itacov/constant/typography.dart';
+import 'package:itacov/ui/constant/constant.dart';
+import 'package:itacov/ui/constant/typography.dart';
+
 
 class BodyCustomWidget extends StatelessWidget {
   @override
